assignment-8/pyproject.toml
[project]
name = "python-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flask>=3.0.2",
    "flask-cors>=4.0.0",
    "google-generativeai>=0.2.2",
    "llama-index>=0.12.28",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "scipy>=1.15.2",
    "numpy>=1.21.0",
    "requests>=2.26.0",
    "tqdm>=4.67.1",
    "pydantic>=1.8.2",
    "python-telegram-bot>=20.7",
]
